feat: time zone setting & Update documentation and some variables

+++ docker/start.sh
@@ -1,14 +1,16 @@
-#!/bin/sh
+#!/bin/bash
 
 cd ${WORKDIR}
-if [ ! -s /tmp/requirements.txt.sha256sum ]; then
-     sha256sum requirements.txt > /tmp/requirements.txt.sha256sum
-fi
+
+if [ "${DLIMITER_AUTO_UPDATE}" = "true" ]; then
+    if [ ! -s /tmp/requirements.txt.sha256sum ]; then
+        sha256sum requirements.txt > /tmp/requirements.txt.sha256sum
+    fi
     echo "更新程序..."
-    git remote set-url origin "${REPO_URL}" &> /dev/null
+    git remote set-url origin "${DLIMITER_REPO_URL}" &> /dev/null
     find . | grep -E "(/__pycache__$|\.pyc$|\.pyo$)" | xargs rm -rf
-    git fetch --depth 1 origin ${BRANCH}
-    git reset --hard origin/${BRANCH}
+    git fetch --depth 1 origin ${DLIMITER_BRANCH}
+    git reset --hard origin/${DLIMITER_BRANCH}
     if [ $? -eq 0 ]; then
         echo "更新成功..."
         # Python依赖包更新
@@ -16,7 +18,7 @@ fi
         hash_new=$(sha256sum requirements.txt)
         if [ "${hash_old}" != "${hash_new}" ]; then
             echo "检测到requirements.txt有变化，重新安装依赖..."
-            if [ "${CN_UPDATE}" = "true" ]; then
+            if [ "${DLIMITER_CN_UPDATE}" = "true" ]; then
                 pip install --upgrade pip wheel -i https://pypi.tuna.tsinghua.edu.cn/simple
                 pip install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple
             else
@@ -33,4 +35,8 @@ fi
     else
         echo "更新失败，继续使用旧的程序来启动..."
     fi
+else
+    echo "程序自动升级已关闭，如需自动升级请在创建容器时设置环境变量：DLIMITER_AUTO_UPDATE=true"
+fi
+
 hypercorn -b 0.0.0.0:8088 main:app
